Updated css ' names to match the components they are scoped for.
Split Body component into more scoped comps. Split css too. Updated css for "QuizStart " component
