+++ my-project/src/components/Footer.module.css
