+++ my-project/src/components/MainContent.module.css
@@ -1,0 +1,9 @@
+.quizContainer {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  overflow: auto;
+  gap: 10px;
+}
