+++ my-project/src/components/QuizStart.module.css
@@ -1,0 +1,62 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+
+  background-color: #2f6370;
+  padding: 40px;
+  border-radius: 1rem;
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
+  width: 500px;
+  margin: auto;
+}
+
+.categoryDropdown {
+  font-size: 16px;
+  padding: 10px;
+  border-radius: 10px;
+  margin-bottom: 20px;
+  width: 300px;
+  background-color: #3C3F58;
+  color: #fefae0;
+  border: 1px solid #fefae0;
+}
+
+.difficultySelector {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 8px;
+  margin: 20px 0;
+  color: #fefae0;
+}
+
+.difficultyLabel {
+  font-weight: bold;
+  color: #fefae0;
+}
+
+.startQuizButton {
+  color: #fefae0;
+  background-color: #3C3F58;
+  border-radius: 15px;
+  padding: 10px 30px;
+  margin: 0 10px;
+  text-align: center;
+  border-color: #fefae0;
+  transition-duration: 0.3s;
+  width: 200px;
+}
+
+.startQuizButton:hover {
+  box-shadow: 0 2px 6px 0 #3c3f5877, 0 7px 15px 0 #54587c65;
+  background-color: #5A7D9A;
+  cursor: pointer;
+}
+
+.startQuizButton:active {
+  transform: translateY(4px);
+  background-color: #3C3F58;
+  cursor: wait;
+}
