--- my-project/src/components/bodyStyle.module.css
@@ -1,137 +1,0 @@
-.main-content {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    overflow: auto;
-    gap: 10px;
-}
-
-.button-start {
-    color: #fefae0;
-    background-color: #3C3F58;
-    border-radius: 15px;
-    padding: 10px;
-    padding-right: 30px;
-    padding-left: 30px;
-    margin-right: 10px;
-    margin-left: 10px;
-    text-align: center;
-    border-color: #fefae0;
-    transition-duration: 0.3s;
-    width: 200px;
-}
-
-.button-start:hover {
-    box-shadow: 0 2px 6px 0 #3c3f5877, 0 7px 15px 0 #54587c65;
-    background-color: #5A7D9A;
-    cursor: pointer;
-}
-
-.button-start:active {
-    transform: translateY(4px);
-    background-color: #3C3F58;
-    cursor: wait;
-}
-
-.question {
-    color: #fefae0;
-    background-color: #2f6370;
-    padding: 30px;
-    border-radius: 1rem;
-    display: flex;
-    flex-direction: column;
-    width: 1000px;
-    text-align: center;
-    font-size: x-large;
-}
-
-.score {
-    color: #fefae0;
-    background-color: #3C3F58;
-    margin: 20px;
-    padding: 20px;
-    border-radius: 1rem;
-    font-size: x-large;
-    font-weight: bold;
-    width: 500px;
-    text-align: center;
-}
-
-.container-multi-answer {
-    background-color: #2f6370;
-    text-align: center;
-    padding: 30px;
-    border-radius: 1rem;
-    display: grid;
-    grid-template-columns: auto auto;
-    row-gap: 20px;
-    width: 1000px;
-}
-
-.container-bool-answer {
-    background-color: #2f6370;
-    text-align: center;
-    padding: 30px;
-    border-radius: 1rem;
-    width: 1000px;
-}
-
-.button-answer {
-    color: #fefae0;
-    background-color: #3C3F58;
-    border-radius: 15px;
-    padding: 10px;
-    padding-right: 30px;
-    padding-left: 30px;
-    margin-right: 10px;
-    margin-left: 10px;
-    text-align: center;
-    border-color: #fefae0;
-    transition-duration: 0.3s;
-    width: 480px;
-}
-
-.button-answer:hover {
-    box-shadow: 0 2px 6px 0 #3c3f5877, 0 7px 15px 0 #54587c65;
-    background-color: #5A7D9A;
-    cursor: pointer;
-}
-
-.button-answer:active {
-    transform: translateY(4px);
-    background-color: #3C3F58;
-}
-
-.end-screen {
-    color: #fefae0;
-    background-color: #2f6370;
-    text-align: center;
-    padding: 30px;
-    border-radius: 1rem;
-    width: 500px;
-    font-size: x-large;
-}
-
-.end-screen p {
-    margin-top: -5px;
-    color: #3C3F58;
-    background-color: #fefae0;
-    border-radius: 1rem;
-}
-
-.easy {
-    color: #fefae0;
-    background-color: #2f7055;
-}
-
-.medium {
-    color: #fefae0;
-    background-color: #774223;
-}
-
-.hard {
-    color: #fefae0;
-    background-color: #832828;
-}
