+++ my-project/src/components/styles/Shared.module.css
@@ -1,0 +1,14 @@
+.easy {
+  color: #fefae0;
+  background-color: #2f7055;
+}
+
+.medium {
+  color: #fefae0;
+  background-color: #774223;
+}
+
+.hard {
+  color: #fefae0;
+  background-color: #832828;
+}
